refactor(sindi): only keep sq8 quantization to term_datas_

+++ src/quantization/sparse_quantization/sparse_term_computer_test.cpp
@@ -69,12 +69,12 @@ TEST_CASE("SparseTermComputer Basic Test", "[ut][SparseTermComputer]") {
     REQUIRE(std::abs(query_val - (-1.0 * query_id)) < 1e-3);
     std::vector<float> dists(10, 0);
     std::vector<uint16_t> term_ids = {0, 2, 4, 6, 8};
-    std::vector<float> term_vals = {0, 2, 4, 6, 8};
+    std::vector<uint8_t> term_vals = {0, 2, 4, 6, 8};
     computer->ScanForAccumulate(
         test_term_it, term_ids.data(), term_vals.data(), term_ids.size(), dists.data());
     for (auto i = 0; i < term_ids.size(); i++) {
         auto id = term_ids[i];
-        REQUIRE(dists[id] == term_vals[i] * query_val);
+        REQUIRE(dists[id] == (float)term_vals[i] * query_val);
     }
 
     // clean
